3 need media locking functionality

+++ Archive.cpp
@@ -246,11 +246,13 @@ namespace arc {
     cold_arc::Result<> Archive::remove(uint64_t id, const std::vector<uint64_t>& media_ids) {
         try {
             std::set<uint64_t> media_set(media_ids.begin(), media_ids.end());
+            auto ref = 0;
             {
                 *m_dbhandle
                     << "SELECT siz, arc_media_id FROM arc_tree INNER JOIN arc_tree_to_media ON (arc_tree.id=arc_tree_to_media.arc_tree_id) WHERE arc_tree.id=?"
                     << id
                     >> [&](sqlite3_uint64 siz, sqlite3_uint64 mid) {
+                        ref++;
                         if (media_set.count(mid)) {
                             {
                                 *m_dbhandle << "UPDATE arc_media SET occupied=occupied-? WHERE id=?" << siz << mid;
@@ -260,10 +262,11 @@ namespace arc {
                                     << "DELETE FROM arc_tree_to_media WHERE arc_tree_id=? AND arc_media_id=?"
                                     << id << mid;
                             }
+                            ref--;
                         }
                     };
             }
-            {
+            if (ref == 0) {
                 *m_dbhandle
                     << "DELETE FROM arc_tree WHERE id=?"
                     << id;
@@ -370,6 +373,16 @@ namespace arc {
             return unexpected_nested(cold_arc::ErrorCode::SettingsSwitchMediaError, res.error());
         m_current_media = std::move(m);
         Signals::instance().update_media_view.emit();
+        return {};
+    }
+    cold_arc::Result<> Archive::Settings::reloadMedia() {
+        if (!m_current_media)
+            return {};
+
+        auto m = std::make_unique<Media>();
+        if (auto res = m->construct(m_dbhandle, m_current_media->m_id); !res)
+            return unexpected_nested(cold_arc::ErrorCode::ReloadMediaError, res.error());
+        m_current_media = std::move(m);
         return {};
     }
 
@@ -485,5 +498,21 @@ namespace arc {
     }
     bool Archive::Media::joliet() const {
         return m_joliet;
+    }
+    cold_arc::Result<> Archive::Media::lock(bool locked) {
+        try {
+            {
+                *m_dbhandle
+                    << "UPDATE arc_media SET locked=? WHERE id=?"
+                    << (locked ? 1 : 0)
+                    << m_id;
+            }
+        } catch (const sqlite::sqlite_exception& e) {
+            return unexpected_sqlite_exception(cold_arc::ErrorCode::MediaLockError, e.get_code());
+        }
+        return {};
+    }
+    bool Archive::Media::locked() const {
+        return m_locked;
     }
 } // arc
